refactor(docs): extract hooks + MCP reference into lib modules
Pure relocation of the two audit-reference clusters out of
scripts/docs/generate.ts:

- lib/hooks-reference.ts: HookRegistration type, hookGroupDescriptions,
  hookId/hookNamespace, the curated hookExamplePayloads + HookExamplePayload,
  formatHookExampleValue/renderHookExample/renderHookCard, and the two public
  renderers renderHookLlmReference (hooks.md bundle) + renderHookReference
  (hooks.mdx).
- lib/mcp-reference.ts: ExecToolDoc type, mcpGroupDescriptions, the
  mcpToolId/Namespace/Label/ExampleValue helpers, renderMcpExample/
  renderMcpToolCard, and the two public renderers renderMcpLlmReference
  (exec-tools.md bundle) + renderExecToolReference (exec-tools.mdx).

Function bodies byte-for-byte unchanged (incl. the irregular frontmatter
indentation in the two .mdx renderers). generate.ts imports the four public
renderers back; extractHookRegistrations/extractExecTools stay imported for the
coverage report. Dropped the now-unused escapeMarkdownTableCell import.
docs:generate output verified byte-identical.
