+++ scripts/docs/generate.ts
@@ -14,7 +14,7 @@ import { getAllRoutes } from '../../src/core/api-docs'
 import { PERMISSION_DESCRIPTIONS, PermissionSchema } from '../../packages/core/src/plugins/permissions'
 import {
   docsBasePath, writeStableFile, docsPath, escapeHtml,
-  escapeTableCell, escapeMarkdownTableCell, generatedPageNote, flattenObject,
+  escapeTableCell, generatedPageNote, flattenObject,
 } from './lib/doc-utils'
 import { renderCliReference } from './lib/cli-reference'
 import { OpenApiOperation, curlForOperation, renderApiReference } from './lib/api-reference'
@@ -25,6 +25,8 @@ import {
   readOfficialPluginCatalog,
 } from './lib/source-extraction'
 import { readSdkExports, renderSdkReference } from './lib/sdk-reference'
+import { renderHookLlmReference, renderHookReference } from './lib/hooks-reference'
+import { renderExecToolReference, renderMcpLlmReference } from './lib/mcp-reference'
 
 const repoRoot = new URL('../..', import.meta.url).pathname
 const docsRoot = join(repoRoot, 'docs')
@@ -302,373 +304,6 @@ function buildCoverageReport(): Record<string, unknown> {
     },
   }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-type HookRegistration = ReturnType<typeof extractHookRegistrations>[number]
-
-const hookGroupDescriptions: Record<string, string> = {
-  assets: 'Asset hooks expose file, sidecar, variant, and trash helpers for plugins that need to work with Bakin-managed files.',
-  health: 'Health hooks expose registered readiness and diagnostic checks so other surfaces can list or inspect them.',
-  models: 'Model hooks expose the effective model configuration and notify dependent surfaces when runtime model state changes.',
-  tasks: 'Task hooks let plugins enrich task details and react to task lifecycle changes.',
-  team: 'Team hooks expose runtime agent and team metadata for plugins that need agent-aware behavior.',
-  workflows: 'Workflow hooks expose workflow definitions, instances, steps, gates, and notification helpers for task automation.',
-}
-
-function hookId(name: string): string {
-  return name
-    .replace(/[^a-zA-Z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    .toLowerCase()
-}
-
-function hookNamespace(name: string): string {
-  return name.split('.')[0] || 'other'
-}
-
-type HookExamplePayload = Record<string, unknown>
-
-const hookExamplePayloads: Record<string, HookExamplePayload> = {
-  'assets.getAssetTypes': {},
-  'assets.resolveServe': { segments: ['20260401-hero-a1b2c3d4', 'thumb'] },
-  'assets.purgeClipboardForTask': { taskId: 'task-123' },
-  'health.list': {},
-  'health.getCheck': { id: 'runtime' },
-  'models.configChanged': { agentId: 'patch', oldModel: 'gpt-5.4', newModel: 'gpt-5.5' },
-  'models.getEffectiveModel': { agentId: 'patch' },
-  'models.markConfigDirty': {},
-  'models.markRuntimeRestarted': {},
-  'models.getAvailableModels': {},
-  'tasks.statusChanged': { taskId: 'task-123', from: 'doing', to: 'done' },
-  'tasks.enrichDetails': { task: { id: 'task-123', projectId: 'launch-docs' } },
-  'team.list': {},
-  'team.getAgent': { id: 'patch' },
-  'team.getAgentIds': {},
-  'team.resolveProfile': { id: 'patch' },
-  'team.getTeamMembers': { teamId: 'docs' },
-  'team.getAgentTeam': { id: 'patch' },
-  'team.getOrgStructure': {},
-  'workflows.loadInstance': { taskId: 'task-123' },
-  'workflows.saveInstance': { instance: { taskId: 'task-123', workflowId: 'docs-review' } },
-  'workflows.createInstance': { taskId: 'task-123', workflowId: 'docs-review', assignee: 'patch' },
-  'workflows.instances.list': { statusFilter: 'in_progress' },
-  'workflows.getCurrentStep': { taskId: 'task-123', agentId: 'patch' },
-  'workflows.completeStep': { taskId: 'task-123', stepId: 'review', output: { ok: true } },
-  'workflows.matchWorkflow': { title: 'Improve hook docs', description: 'Add generated examples' },
-  'workflows.definitions.list': {},
-  'workflows.loadDefinition': { name: 'docs-review' },
-  'workflows.getActiveAgents': { taskId: 'task-123' },
-  'workflows.isGateNotified': { taskId: 'task-123', stepId: 'approval' },
-  'workflows.markGateNotified': { taskId: 'task-123', stepId: 'approval' },
-  'workflows.validateStepOutput': { schema: { type: 'object' }, output: { approved: true } },
-  'workflows.cancelInstance': { taskId: 'task-123' },
-  'workflows.notificationChannels.list': {},
-  'workflows.getNotificationChannel': { id: 'slack' },
-}
-
-function formatHookExampleValue(value: unknown, indent = 0): string {
-  const pad = ' '.repeat(indent)
-  const childPad = ' '.repeat(indent + 2)
-  if (typeof value === 'string') return `'${value.replace(/'/g, "\\'")}'`
-  if (typeof value === 'number' || typeof value === 'boolean') return String(value)
-  if (Array.isArray(value)) {
-    if (!value.length) return '[]'
-    return `[\n${value.map(item => `${childPad}${formatHookExampleValue(item, indent + 2)}`).join(',\n')}\n${pad}]`
-  }
-  if (value && typeof value === 'object') {
-    const entries = Object.entries(value as Record<string, unknown>)
-    if (!entries.length) return '{}'
-    return `{\n${entries.map(([key, item]) => `${childPad}${key}: ${formatHookExampleValue(item, indent + 2)}`).join(',\n')}\n${pad}}`
-  }
-  return 'null'
-}
-
-function renderHookExample(hook: HookRegistration): string {
-  const payload = formatHookExampleValue(hookExamplePayloads[hook.name] ?? {}, 2)
-  const method = hook.hookKind === 'event'
-    ? 'callAll'
-    : hook.hookKind === 'waterfall'
-      ? 'call'
-      : 'invoke'
-  if (hook.hookKind === 'event') {
-    return `await ctx.hooks.${method}(\n  '${hook.name}',\n  ${payload},\n)`
-  }
-  if (hook.hookKind === 'waterfall') {
-    return `const next = await ctx.hooks.${method}(\n  '${hook.name}',\n  ${payload},\n)`
-  }
-  return `const result = await ctx.hooks.${method}(\n  '${hook.name}',\n  ${payload},\n)`
-}
-
-function renderHookCard(hook: HookRegistration): string {
-  const id = hookId(hook.name)
-  const label = hook.label ?? hook.summary ?? hook.name
-  const badges = [
-    hook.hookKind,
-    hook.visibility && hook.visibility !== 'public' ? hook.visibility : '',
-    hook.stability && hook.stability !== 'stable' ? hook.stability : '',
-  ].filter(Boolean)
-  const props = [
-    `id="${id}"`,
-    `name="${escapeHtml(hook.name)}"`,
-    `label="${escapeHtml(label)}"`,
-    hook.summary ? `summary="${escapeHtml(hook.summary)}"` : '',
-    `source="${escapeHtml(`${hook.file}:${hook.line}`)}"`,
-    badges.length ? `badges={${JSON.stringify(badges)}}` : '',
-  ].filter(Boolean).join(' ')
-  return [
-    `<HookCard ${props}>`,
-    '```ts frame="terminal"',
-    renderHookExample(hook),
-    '```',
-    '</HookCard>',
-  ].filter(Boolean).join('\n')
-}
-
-function renderHookLlmReference(): string {
-  const hooks = extractHookRegistrations()
-  const grouped = new Map<string, HookRegistration[]>()
-  for (const hook of hooks) {
-    const namespace = hookNamespace(hook.name)
-    const existing = grouped.get(namespace) ?? []
-    existing.push(hook)
-    grouped.set(namespace, existing)
-  }
-
-  const lines = [
-    'Hooks are plugin integration points. Use them from plugin code through `ctx.hooks`.',
-    '',
-    'Invocation style depends on `kind`:',
-    '',
-    '- `rpc`: `await ctx.hooks.invoke(name, payload)`',
-    '- `event`: `await ctx.hooks.callAll(name, payload)`',
-    '- `waterfall`: `await ctx.hooks.call(name, payload)`',
-    '',
-  ]
-
-  for (const namespace of [...grouped.keys()].sort((a, b) => a.localeCompare(b))) {
-    const title = namespace === 'other' ? 'Other' : `${namespace[0].toUpperCase()}${namespace.slice(1)}`
-    lines.push(`## ${title}`, '')
-    const description = hookGroupDescriptions[namespace]
-    if (description) lines.push(description, '')
-    for (const hook of [...(grouped.get(namespace) ?? [])].sort((a, b) => a.name.localeCompare(b.name))) {
-      lines.push(`### ${hook.name}`, '')
-      lines.push(`Label: ${hook.label ?? hook.summary ?? hook.name}`)
-      if (hook.summary) lines.push(`Purpose: ${hook.summary}`)
-      lines.push(`Kind: ${hook.hookKind ?? 'rpc'}`)
-      lines.push(`Source: ${hook.file}:${hook.line}`)
-      lines.push('', 'Example:', '', '```ts', renderHookExample(hook), '```', '')
-    }
-  }
-
-  return lines.join('\n')
-}
-
-type ExecToolDoc = ReturnType<typeof extractExecTools>[number]
-
-const mcpGroupDescriptions: Record<string, string> = {
-  assets: 'Asset tools let agents list, inspect, save, link, restore, and clean up files managed by Bakin.',
-  gen: 'Generation tools create or import media through Bakin so outputs land in the asset pipeline with task context.',
-  get: 'Runtime lookup tools return local Bakin paths and state agents should not hardcode.',
-  health: 'Health tools let agents check whether Bakin is running correctly before or during work.',
-  heartbeat: 'Heartbeat tools let agents publish lightweight status so Bakin can show who is active and what they are doing.',
-  log: 'Logging tools record progress updates in Bakin task history and audit surfaces.',
-  memory: 'Memory tools expose indexed runtime memory, sessions, turns, checkpoints, and status to agents.',
-  messaging: 'Messaging tools let agents create, update, approve, reject, and inspect human-facing messages and sessions.',
-  models: 'Model tools expose model configuration and available model choices to agents.',
-  post: 'Publishing tools send completed work to configured channels through Bakin adapters.',
-  project: 'Project tools let agents create, update, read, and maintain project specs and linked checklist items.',
-  schedule: 'Schedule tools let agents create, inspect, pause, run, and update recurring Bakin jobs.',
-  search: 'Search tools query Bakin-indexed content across plugins or inside a specific surface.',
-  submit: 'Workflow submission tools let agents submit step output back to the workflow engine.',
-  tasks: 'Task tools are the main agent interface for creating, reading, moving, logging, blocking, and completing work.',
-  team: 'Team tools expose agent roster, identity, status, messaging, and permission operations.',
-  workflows: 'Workflow tools expose workflow definitions, active instances, current steps, and step completion.',
-}
-
-function mcpToolId(name: string): string {
-  return name.replace(/^bakin_exec_/, '').replace(/[^a-zA-Z0-9]+/g, '-').replace(/^-+|-+$/g, '').toLowerCase()
-}
-
-function mcpToolNamespace(name: string): string {
-  return name.replace(/^bakin_exec_/, '').split('_')[0] || 'other'
-}
-
-function mcpToolLabel(name: string): string {
-  const words = name.replace(/^bakin_exec_/, '').split('_').filter(Boolean)
-  return `${words.map(word => word[0]?.toUpperCase() + word.slice(1)).join(' ')}.`
-}
-
-function mcpExampleValue(type: string): unknown {
-  if (type === 'number') return 20
-  if (type === 'boolean') return true
-  if (type === 'choice') return 'value'
-  if (type === 'record' || type === 'object') return { key: 'value' }
-  if (type === 'array') return ['value']
-  return 'value'
-}
-
-function renderMcpExample(tool: ExecToolDoc): string {
-  const args = Object.fromEntries(tool.parameters.map(param => [param.name, mcpExampleValue(param.type)]))
-  const json = JSON.stringify(args, null, 2)
-  if (tool.parameters.length === 0) return `mcporter call bakin-<agent>.${tool.name}`
-  return `mcporter call bakin-<agent>.${tool.name} --args '${json}'`
-}
-
-function renderMcpToolCard(tool: ExecToolDoc): string {
-  const id = mcpToolId(tool.name)
-  const props = [
-    `id="${id}"`,
-    `name="${escapeHtml(tool.name)}"`,
-    `label="${escapeHtml(tool.label ?? mcpToolLabel(tool.name))}"`,
-    tool.description ? `description="${escapeHtml(tool.description.trim())}"` : '',
-    `source="${escapeHtml(`${tool.file}:${tool.line}`)}"`,
-    tool.parameters.length ? `parameters={${JSON.stringify(tool.parameters)}}` : '',
-  ].filter(Boolean).join(' ')
-  return [
-    `<McpToolCard ${props}>`,
-    '```sh frame="terminal"',
-    renderMcpExample(tool),
-    '```',
-    '</McpToolCard>',
-  ].join('\n')
-}
-
-function renderMcpLlmReference(): string {
-  const tools = extractExecTools()
-  const grouped = new Map<string, ExecToolDoc[]>()
-  for (const tool of tools) {
-    const namespace = mcpToolNamespace(tool.name)
-    const existing = grouped.get(namespace) ?? []
-    existing.push(tool)
-    grouped.set(namespace, existing)
-  }
-
-  const lines = [
-    'Use MCP tools through `mcporter`:',
-    '',
-    '```sh',
-    "mcporter call bakin-<agent>.<tool_name> --args '<json>'",
-    '```',
-    '',
-    'Use the exact tool name shown below. Omit `--args` only for tools with no parameters.',
-    '',
-  ]
-
-  for (const namespace of [...grouped.keys()].sort((a, b) => a.localeCompare(b))) {
-    const title = namespace === 'other' ? 'Other' : `${namespace[0].toUpperCase()}${namespace.slice(1)}`
-    lines.push(`## ${title}`, '')
-    const description = mcpGroupDescriptions[namespace]
-    if (description) lines.push(description, '')
-    for (const tool of [...(grouped.get(namespace) ?? [])].sort((a, b) => a.name.localeCompare(b.name))) {
-      lines.push(`### ${tool.name}`, '')
-      lines.push(`Label: ${tool.label ?? mcpToolLabel(tool.name)}`)
-      if (tool.description) lines.push(`Purpose: ${tool.description.trim()}`)
-      if (tool.parameters.length) {
-        lines.push('', '| Argument | Type | Required | Description |', '| --- | --- | --- | --- |')
-        for (const param of tool.parameters) {
-          lines.push(`| \`${escapeMarkdownTableCell(param.name)}\` | ${escapeMarkdownTableCell(param.type)} | ${param.required ? 'yes' : 'no'} | ${escapeMarkdownTableCell(param.description ?? '')} |`)
-        }
-      } else {
-        lines.push('', 'Arguments: none.')
-      }
-      lines.push('', 'Example:', '', '```sh', renderMcpExample(tool), '```', '')
-    }
-  }
-
-  return lines.join('\n')
-}
-
-function renderHookReference(): string {
-  const hooks = extractHookRegistrations()
-  const grouped = new Map<string, HookRegistration[]>()
-  for (const hook of hooks) {
-    const namespace = hookNamespace(hook.name)
-    const existing = grouped.get(namespace) ?? []
-    existing.push(hook)
-    grouped.set(namespace, existing)
-  }
-
-  const lines = [
-    '---',
-    'title: Hooks',
-    'description: Generated audit reference for Bakin hook registrations.',
-	    '---',
-	    '',
-    "import HookCard from '../../../../components/HookCard.astro'",
-    '',
-    '<div class="hook-reference-intro">',
-    '  <p>Hooks are the integration points plugins use to share Bakin state and behavior. Reach for them when you need task context, agent details, model choices, workflow state, assets, or health checks owned by another plugin.</p>',
-    '  <p>This reference shows the hook key to call, what it returns or changes, and the registration that owns the contract.</p>',
-    '</div>',
-    '',
-	  ]
-
-  for (const namespace of [...grouped.keys()].sort((a, b) => a.localeCompare(b))) {
-    const namespaceHooks = [...(grouped.get(namespace) ?? [])].sort((a, b) => a.name.localeCompare(b.name))
-    const title = namespace === 'other' ? 'Other' : `${namespace[0].toUpperCase()}${namespace.slice(1)}`
-    lines.push(`## ${title}`, '')
-    lines.push(`<p class="hook-section-description">${escapeHtml(hookGroupDescriptions[namespace] ?? 'Hooks discovered from source registration audit.')}</p>`, '')
-    lines.push('<div class="hook-card-list">')
-    for (const hook of namespaceHooks) {
-      lines.push(renderHookCard(hook))
-    }
-    lines.push('</div>', '')
-	  }
-	  lines.push(generatedPageNote(), '')
-
-	  return lines.join('\n')
-	}
-
-function renderExecToolReference(): string {
-  const tools = extractExecTools()
-  const grouped = new Map<string, ExecToolDoc[]>()
-  for (const tool of tools) {
-    const namespace = mcpToolNamespace(tool.name)
-    const existing = grouped.get(namespace) ?? []
-    existing.push(tool)
-    grouped.set(namespace, existing)
-  }
-
-  const lines = [
-    '---',
-    'title: MCP',
-    'description: Generated reference for Bakin MCP tools exposed to agents.',
-	    '---',
-	    '',
-    "import McpToolCard from '../../../../components/McpToolCard.astro'",
-    '',
-    '<div class="mcp-reference-intro">',
-    '  <p>MCP tools are how agents get work done in Bakin. They create and move tasks, advance workflows, manage assets and projects, search local state, send messages, schedule recurring work, and check system health.</p>',
-    '  <p>Use this reference when you need the exact tool name, the arguments it accepts, and a copyable call shape.</p>',
-    '</div>',
-    '',
-	  ]
-
-  for (const namespace of [...grouped.keys()].sort((a, b) => a.localeCompare(b))) {
-    const namespaceTools = [...(grouped.get(namespace) ?? [])].sort((a, b) => a.name.localeCompare(b.name))
-    const title = namespace === 'other' ? 'Other' : `${namespace[0].toUpperCase()}${namespace.slice(1)}`
-    lines.push(`## ${title}`, '')
-    lines.push(`<p class="mcp-section-description">${escapeHtml(mcpGroupDescriptions[namespace] ?? 'MCP tools discovered from registered exec tool definitions.')}</p>`, '')
-    lines.push('<div class="mcp-tool-list">')
-    for (const tool of namespaceTools) {
-      lines.push(renderMcpToolCard(tool))
-    }
-    lines.push('</div>', '')
-	  }
-	  lines.push(generatedPageNote(), '')
-
-	  return lines.join('\n')
-	}
 
 function renderPluginCatalog(): string {
   const plugins = readOfficialPluginCatalog()
